Add files via upload
This is my "Seize the target" boat code that is uses IMU, fans, and MQTT to navigate toward a target. It connects to WiFi, gets position data from ArUco markers, and uses PID control to adjust direction and speed. The boat keeps moving and correcting its course until it reaches the target, showing updates on an OLED screen.