Water description(hover over)
This program automates plant care using an ESP32 microcontroller with sensors and a water pump. It monitors soil moisture, temperature, humidity, and light intensity, and sends the data to an MQTT server for tracking. The system reads soil moisture levels and waters the plant if the soil is too dry or if 24 hours have passed since the last watering. Depending on the soil moisture sensor, it will either pump 20 mL of water or 50 mL if the humidity of the soil is below 40%. Temperature and humidity are measured with a DHT sensor, while light intensity is calculated from a photocell reading. In the code there is an equation that converts the volts measured from the photocell to lux. All sensor data is published to specific MQTT topics every 10 minutes and it creates a graph of the data on the sit. It is also printed to the serial monitor for debugging. This setup combines autonomy, efficiency, and smart resource management for a smart watering system.